requirements.txt
Flask>=3.0.0
werkzeug>=3.0.0
pystray>=0.19.5
Pillow>=10.0.0
waitress>=3.0.0
paramiko>=3.4.0
olefile>=0.47
